requirements.txt
asttokens==3.0.0
decorator==5.2.1
executing==2.2.1
iniconfig==2.3.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.3.4
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pygame-ce==2.5.6
Pygments==2.19.2
pytest==8.4.2
ruff==0.14.1
setuptools==80.9.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.14